fixed the json bug and added a safety if a document was not a josn nor a csv
