+++ V2/Cargohub/Program.cs
@@ -21,6 +21,7 @@ builder.Services.AddTransient<IShipmentService, ShipmentService>();
 builder.Services.AddTransient<ISupplierService, SupplierService>();
 builder.Services.AddTransient<ITransferService, TransferService>();
 builder.Services.AddTransient<IOrderService, OrderService>();
+builder.Services.AddTransient<IAdminService, AdminService>();
 
 var app = builder.Build();
