+++ V2/Cargohub/services/Iadminservice.cs
@@ -2,5 +2,5 @@ namespace ServicesV2;
 
 public interface IAdminService
 {
-
+    string AddData(IFormFile file);
 }
